Add types.ts for sharedLink and permission objects for one-drive integration
